Broadcast , memegen , etc fix

+++ commands/group.js
@@ -235,16 +235,17 @@ cmd({
     )
     //---------------------------------------------------------------------------
 cmd({
-            pattern: "resetwarn",
-            desc: "Deletes all previously given warns to quoted user.",
+            pattern: "rwarn",
+            desc: "Deletes all previously given warns of quoted user.",
             category: "group",
-           filename: __filename,
+            filename: __filename,
             use: '<quote|reply|number>',
         },
-        async(Void, citel, text) => {
+        async(Void, citel, text,{isCreator}) => {
             if (!isCreator) return citel.reply(tlang().owner)
+            if (!citel.quoted) return citel.reply('Quote a user master.')
             await warndb.deleteOne({ id: citel.quoted.sender.split('@')[0] + 'warn' });
-            citel.reply('User is free as a bird.\nAll previously given warn has been deleted.')
+            return citel.reply('User is now free as a bird.\n.')
         }
     )
     //---------------------------------------------------------------------------
@@ -593,9 +594,17 @@ cmd({
             if (!/image/.test(mime)) return citel.reply(`Reply to Photo With Caption *text*`)
             mee = await Void.downloadAndSaveMediaMessage(citel.quoted)
             mem = await TelegraPh(mee)
-            meme = `https://api.memegen.link/images/custom/-/${text}.png?background=${mem}`
-            memek = await Void.sendImageAsSticker(citel.chat, meme, citel, { packname: citel.pushName, author: 'khadherinc' })
-            await fs.unlinkSync(memek)
+            meme = await getBuffer(`https://api.memegen.link/images/custom/-/${text}.png?background=${mem}`)
+            let buttonMessage = {
+                image: meme,
+                caption: "Here we go",
+                footer: tlang().footer,
+                headerType: 4,
+            };
+            Void.sendMessage(citel.chat, buttonMessage, {
+                quoted: citel,
+            });
+            await fs.unlinkSync(mee)
 
         }
     )
@@ -901,6 +910,7 @@ cmd({
 )
 
 //---------------------------------------------------------------------------
+if(Config.WORKTYPE!=='private'){
 cmd({ on: "text" }, async(Void, citel) => {
     const randomXp = 8;
     let usrname = citel.pushName
@@ -983,3 +993,4 @@ cmd({ on: "text" }, async(Void, citel) => {
     }
 
 })
+}
